as of 27-4

+++ app/Controllers/Http/PostController.js
@@ -1,10 +1,11 @@
 'use strict'
 const Post = use('App/Models/Post');
 const { validate } = use('Validator');
+const uuid = use('uuid');
 
 class PostController {
     
-    async index ({ request, response, view }) {
+    async index ({ view }) {
         
         const posts = await Post.all();
 
@@ -12,6 +13,15 @@ class PostController {
             title: 'Check out the Posts',
             posts: posts.toJSON()
         });
+    }
+
+    async myIndex ({ view, auth }) {
+        const posts = await Post.query().where('user_id', auth.user.id).fetch();
+
+        return view.render('posts.myIndex', {
+            title: `Posts made by ${auth.user.username}`,
+            posts: posts.toJSON()
+        })
     }
 
     async details ({ params, view }) {
@@ -29,7 +39,7 @@ class PostController {
         })
     }
 
-    async store ({ request, response, session }) {
+    async store ({ request, auth, response, session }) {
         const post = new Post();
         const rules = {
             title: 'required|min:3|max:255',
@@ -43,9 +53,11 @@ class PostController {
             return response.redirect('back');
         }
 
+        post.primaryKeyValue = uuid.v4();
         post.title = request.input('title');
         post.body = request.input('body');
-
+        
+        post.user_id = auth.user.id;
 
 
         await post.save();
@@ -55,17 +67,26 @@ class PostController {
         return response.redirect('/posts');
     }
 
-    async edit ({ params, view }) {
+    async edit ({ params, view, auth, response }) {
         const post = await Post.find(params.id);
 
-        return view.render('/posts.edit', {
-            title: 'Edit Post',
-            post
-        })
+        if (auth.user.id === post.user_id){
+            return view.render('/posts.edit', {
+                title: 'Edit Post',
+                post
+            })
+        }
+
+        return response.redirect('/posts');
     }
 
-    async update ({ params, request, response, session}) {
+    async update ({ params, request, response, session, auth}) {
         const post = await Post.find(params.id);
+
+        if (auth.user.id !== post.user_id) {
+            return response.redirect('/posts');
+        }
+
         const rules = {
             title: 'required|min:3|max:255',
             body: 'required|min:3'
@@ -90,12 +111,13 @@ class PostController {
         return response.redirect('/posts');
     }
 
-    async destroy ({params, response, session }) {
+    async destroy ({params, response, session, auth }) {
         const post = await Post.find(params.id);
-        
-        await post.delete();
 
-        session.flash({ notification: 'Post deleted.'});
+        if (auth.user.id === post.user_id){
+            await post.delete();
+            session.flash({ notification: 'Post deleted.'});
+        }
 
         return response.redirect('/posts');
     }
